requirements.txt
Flask==2.3.3
pymongo==4.8.0
opencv-python==4.9.0.80
python-dotenv==1.0.1
face_recognition==1.3.0
cmake==3.27.0
numpy>=1.24.3
ultralytics==8.1.0